add resource cells spawn frequency

+++ sources/perlin_noise/perlin_noise.c
@@ -31,34 +31,34 @@ int noise2(int x, int y) {
     return permutation[(tmp + x) % 256];
 }
 
-float lin_inter(float x, float y, float s) {
+double lin_inter(double x, double y, double s) {
     return x + s * (y-x);
 }
 
-float smooth_inter(float x, float y, float s) {
+double smooth_inter(double x, double y, double s) {
     return lin_inter(x, y, s*s * (3 - 2*s));
 }
 
-float noise2d(float x, float y) {
+double noise2d(double x, double y) {
     int x_int = x;
     int y_int = y;
-    float x_frac = x - x_int;
-    float y_frac = y - y_int;
+    double x_frac = x - x_int;
+    double y_frac = y - y_int;
     int s = noise2(x_int, y_int);
     int t = noise2(x_int+1, y_int);
     int u = noise2(x_int, y_int+1);
     int v = noise2(x_int+1, y_int+1);
-    float low = smooth_inter(s, t, x_frac);
-    float high = smooth_inter(u, v, x_frac);
+    double low = smooth_inter(s, t, x_frac);
+    double high = smooth_inter(u, v, x_frac);
     return smooth_inter(low, high, y_frac);
 }
 
-float perlin(float x, float y, float freq, int octave) {
-    float xa = x*freq;
-    float ya = y*freq;
-    float amp = 1.0;
-    float fin = 0;
-    float div = 0.0;
+double perlin(double x, double y, double freq, int octave) {
+    double xa = x*freq;
+    double ya = y*freq;
+    double amp = 1.0;
+    double fin = 0;
+    double div = 0.0;
 
     int i;
     for(i=0; i<octave; i++) {
